@@ -49,7 +49,7 @@ export default class Modal extends React.Component<
 
     private layer: ReturnType<typeof LayerManager.renderLayer>;
 
-    private closeCallback: (resp?: any) => void;
+    private closeCallback: (resp?: unknown) => void;
 
     private onClose = () => {
         this.setState({
