@@ -1,6 +1,4 @@
-import React from 'react';
-
-export default function FiberManualRecord(props: any) {
+export default function FiberManualRecord(props: object) {
     return (
         <svg height="24" viewBox="0 0 24 24" width="24" fill="currentColor" {...props}>
             <path d="M24 24H0V0h24v24z" fill="none" />
