@@ -1,13 +1,17 @@
-import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
-import { Accordion, AccordionStep, AccordionStepBody, AccordionStepFooter } from '../../../src/components/Accordion';
+import {
+    Accordion,
+    AccordionStep,
+    AccordionStepBody,
+    AccordionStepFooter,
+} from '../../../src/components/Accordion';
 import { Button, ActionButton } from '../../../src/components/Button';
 import { Dummy } from '../../resources/Dummy';
 
 const meta: Meta<typeof Accordion> = {
-  component: Accordion,
-  title: 'Declarative Components/Accordion',
-  tags: ['!dev', '!autodocs'],
+    component: Accordion,
+    title: 'Declarative Components/Accordion',
+    tags: ['!dev', '!autodocs'],
 };
 
 export default meta;
@@ -16,8 +20,7 @@ type Story = StoryObj;
 export const Default: Story = {
     render: () => (
         <Accordion>
-            {/* @ts-ignore - Accordion injects onStepClick and open props */}
-            <AccordionStep header='Welcome' completed>
+            <AccordionStep header="Welcome" completed>
                 <AccordionStepBody>
                     <Dummy>Dummy content for Welcome step</Dummy>
                 </AccordionStepBody>
@@ -26,16 +29,13 @@ export const Default: Story = {
                     <ActionButton>Next</ActionButton>
                 </AccordionStepFooter>
             </AccordionStep>
-            {/* @ts-ignore */}
-            <AccordionStep header='Basic Details' completed>
+            <AccordionStep header="Basic Details" completed>
                 <Dummy>Dummy content for Basic Details step</Dummy>
             </AccordionStep>
-            {/* @ts-ignore */}
-            <AccordionStep header='Personalisation' errorText='3 Errors' disabled>
+            <AccordionStep header="Personalisation" errorText="3 Errors" disabled>
                 <Dummy>Dummy content for Personalisation step</Dummy>
             </AccordionStep>
-            {/* @ts-ignore */}
-            <AccordionStep header='Verification'>
+            <AccordionStep header="Verification">
                 <Dummy>Dummy content for Verification step</Dummy>
             </AccordionStep>
         </Accordion>
@@ -62,15 +62,19 @@ export const Default: Story = {
     <AccordionStep header='Verification'>
         <Dummy>Dummy content for Verification step</Dummy>
     </AccordionStep>
-</Accordion>`
-            }
-        }
-    }
+</Accordion>`,
+            },
+        },
+    },
 };
 
 // Props documentation story
 export const AccordionProps: Story = {
-    render: () => <Accordion><div /></Accordion>,
+    render: () => (
+        <Accordion>
+            <div />
+        </Accordion>
+    ),
     args: {},
 };
 
