@@ -1,11 +1,10 @@
-import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import { AccordionStep } from '../../../src/components/Accordion';
 
 const meta: Meta<typeof AccordionStep> = {
-  component: AccordionStep as any,
-  title: 'Declarative Components/Accordion/AccordionStep',
-  tags: ['!dev', '!autodocs'],
+    component: AccordionStep as never,
+    title: 'Declarative Components/Accordion/AccordionStep',
+    tags: ['!dev', '!autodocs'],
 };
 
 export default meta;
