@@ -1,12 +1,17 @@
-import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
-import { Button, LinkButton, RaisedButton, ActionButton, IconButton } from '../../../src/components/Button';
+import {
+    Button,
+    LinkButton,
+    RaisedButton,
+    ActionButton,
+    IconButton,
+} from '../../../src/components/Button';
 import SearchIcon from '../../resources/SearchIcon';
 
 const meta: Meta<typeof Button> = {
-  component: Button,
-  title: 'Declarative Components/Button',
-  tags: ['!dev', '!autodocs'],
+    component: Button,
+    title: 'Declarative Components/Button',
+    tags: ['!dev', '!autodocs'],
 };
 
 export default meta;
@@ -15,11 +20,21 @@ type Story = StoryObj;
 export const EnabledState: Story = {
     render: () => (
         <div style={{ display: 'flex', gap: '8px', flexWrap: 'wrap' }}>
-            <Button><SearchIcon /> Button</Button>
-            <LinkButton><SearchIcon /> Link Button</LinkButton>
-            <RaisedButton><SearchIcon /> Raised Button</RaisedButton>
-            <ActionButton><SearchIcon /> Action Button</ActionButton>
-            <IconButton><SearchIcon /></IconButton>
+            <Button>
+                <SearchIcon /> Button
+            </Button>
+            <LinkButton>
+                <SearchIcon /> Link Button
+            </LinkButton>
+            <RaisedButton>
+                <SearchIcon /> Raised Button
+            </RaisedButton>
+            <ActionButton>
+                <SearchIcon /> Action Button
+            </ActionButton>
+            <IconButton>
+                <SearchIcon />
+            </IconButton>
         </div>
     ),
     parameters: {
@@ -29,20 +44,30 @@ export const EnabledState: Story = {
 <LinkButton><SearchIcon /> Link Button</LinkButton>
 <RaisedButton><SearchIcon /> Raised Button</RaisedButton>
 <ActionButton><SearchIcon /> Action Button</ActionButton>
-<IconButton><SearchIcon /></IconButton>`
-            }
-        }
-    }
+<IconButton><SearchIcon /></IconButton>`,
+            },
+        },
+    },
 };
 
 export const DisabledState: Story = {
     render: () => (
         <div style={{ display: 'flex', gap: '8px', flexWrap: 'wrap' }}>
-            <Button disabled><SearchIcon /> Disabled Button</Button>
-            <LinkButton disabled><SearchIcon /> Disabled Link Button</LinkButton>
-            <RaisedButton disabled><SearchIcon /> Disabled Raised Button</RaisedButton>
-            <ActionButton disabled><SearchIcon /> Disabled Action Button</ActionButton>
-            <IconButton disabled><SearchIcon /></IconButton>
+            <Button disabled>
+                <SearchIcon /> Disabled Button
+            </Button>
+            <LinkButton disabled>
+                <SearchIcon /> Disabled Link Button
+            </LinkButton>
+            <RaisedButton disabled>
+                <SearchIcon /> Disabled Raised Button
+            </RaisedButton>
+            <ActionButton disabled>
+                <SearchIcon /> Disabled Action Button
+            </ActionButton>
+            <IconButton disabled>
+                <SearchIcon />
+            </IconButton>
         </div>
     ),
     parameters: {
@@ -52,10 +77,10 @@ export const DisabledState: Story = {
 <LinkButton disabled><SearchIcon /> Disabled Link Button</LinkButton>
 <RaisedButton disabled><SearchIcon /> Disabled Raised Button</RaisedButton>
 <ActionButton disabled><SearchIcon /> Disabled Action Button</ActionButton>
-<IconButton disabled><SearchIcon /></IconButton>`
-            }
-        }
-    }
+<IconButton disabled><SearchIcon /></IconButton>`,
+            },
+        },
+    },
 };
 
 // Props documentation story
