@@ -1,33 +1,32 @@
-import React from "react";
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import { ChipInput } from '../../../src/components/ChipInput';
 
 const meta: Meta<typeof ChipInput> = {
-  component: ChipInput as any,
-  title: 'Declarative Components/ChipInput',
-  tags: ['!dev', '!autodocs'],
+    component: ChipInput as never,
+    title: 'Declarative Components/ChipInput',
+    tags: ['!dev', '!autodocs'],
 };
 
 export default meta;
 type Story = StoryObj<typeof ChipInput>;
 
 export const Primary: Story = {
-  args: {
-    label: 'Enter your Skills',
-  },
+    args: {
+        label: 'Enter your Skills',
+    },
 };
 
 export const Variants: Story = {
-  render: () => (
-    <div>
-      <ChipInput label="Enter your Skills" required />
-    </div>
-  ),
-  parameters: {
-    docs: {
-      source: {
-        code: `<ChipInput label="Enter your Skills" />`,
-      },
+    render: () => (
+        <div>
+            <ChipInput label="Enter your Skills" required />
+        </div>
+    ),
+    parameters: {
+        docs: {
+            source: {
+                code: `<ChipInput label="Enter your Skills" />`,
+            },
+        },
     },
-  },
 };
