@@ -1,42 +1,41 @@
-import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import { Card } from '../../../src/components/Card';
 import { Menu, MenuItem } from '../../../src/components/Menu';
 
 const meta: Meta<typeof Menu> = {
-  component: Menu,
-  title: 'Declarative Components/Menu',
-  tags: ['!dev', '!autodocs'],
+    component: Menu,
+    title: 'Declarative Components/Menu',
+    tags: ['!dev', '!autodocs'],
 };
 
 export default meta;
 type Story = StoryObj<typeof Menu>;
 
 export const Variants: Story = {
-  render: () => (
-    <div style={{ display: 'flex', gap: '16px', alignItems: 'flex-start' }}>
-      <Card style={{ width: '200px' }}>
-        <Menu>
-          <MenuItem value='apple'>Apple</MenuItem>
-          <MenuItem value='orange'>Orange</MenuItem>
-          <MenuItem value='mango'>Mango</MenuItem>
-          <MenuItem value='Banana'>Banana</MenuItem>
-        </Menu>
-      </Card>
-      <Card style={{ width: '200px' }}>
-        <Menu multiSelect>
-          <MenuItem value='apple'>Apple</MenuItem>
-          <MenuItem value='orange'>Orange</MenuItem>
-          <MenuItem value='mango'>Mango</MenuItem>
-          <MenuItem value='Banana'>Banana</MenuItem>
-        </Menu>
-      </Card>
-    </div>
-  ),
-  parameters: {
-    docs: {
-      source: {
-        code: `<Card>
+    render: () => (
+        <div style={{ display: 'flex', gap: '16px', alignItems: 'flex-start' }}>
+            <Card style={{ width: '200px' }}>
+                <Menu>
+                    <MenuItem value="apple">Apple</MenuItem>
+                    <MenuItem value="orange">Orange</MenuItem>
+                    <MenuItem value="mango">Mango</MenuItem>
+                    <MenuItem value="Banana">Banana</MenuItem>
+                </Menu>
+            </Card>
+            <Card style={{ width: '200px' }}>
+                <Menu multiSelect>
+                    <MenuItem value="apple">Apple</MenuItem>
+                    <MenuItem value="orange">Orange</MenuItem>
+                    <MenuItem value="mango">Mango</MenuItem>
+                    <MenuItem value="Banana">Banana</MenuItem>
+                </Menu>
+            </Card>
+        </div>
+    ),
+    parameters: {
+        docs: {
+            source: {
+                code: `<Card>
         <Menu>
           <MenuItem value='apple'>Apple</MenuItem>
           <MenuItem value='orange'>Orange</MenuItem>
@@ -52,22 +51,28 @@ export const Variants: Story = {
           <MenuItem value='Banana'>Banana</MenuItem>
         </Menu>
       </Card>`,
-      },
+            },
+        },
     },
-  },
 };
 
 export const MenuProps: Story = {
-  args: {
-    children: [
-      <MenuItem key="1" value='option1'>Option 1</MenuItem>,
-      <MenuItem key="2" value='option2'>Option 2</MenuItem>,
-      <MenuItem key="3" value='option3'>Option 3</MenuItem>,
-    ],
-  },
+    args: {
+        children: [
+            <MenuItem key="1" value="option1">
+                Option 1
+            </MenuItem>,
+            <MenuItem key="2" value="option2">
+                Option 2
+            </MenuItem>,
+            <MenuItem key="3" value="option3">
+                Option 3
+            </MenuItem>,
+        ],
+    },
 };
 
 export const MenuItemProps: Story = {
-  // @ts-ignore - This is for Controls only
-  render: (args) => <MenuItem {...args}>Item</MenuItem>,
+    // @ts-expect-error - This is for Controls only
+    render: (args) => <MenuItem {...args}>Item</MenuItem>,
 };
