@@ -1,6 +1,5 @@
-import React from "react";
 import type { Meta, StoryObj } from '@storybook/react';
-import { AlertDialog } from "../../../src/components";
+import { AlertDialog } from '../../../src/components';
 
 const meta: Meta<typeof AlertDialog> = {
     component: AlertDialog,
@@ -13,5 +12,5 @@ type Story = StoryObj<typeof AlertDialog>;
 
 // Props documentation story for AlertDialog
 export const AlertDialogProps: Story = {
-    render: (args: any) => <AlertDialog {...args} />,
+    render: (args: object) => <AlertDialog {...args} />,
 };
