@@ -1,6 +1,5 @@
-import React from "react";
 import type { Meta, StoryObj } from '@storybook/react';
-import { PromptDialog } from "../../../src/components";
+import { PromptDialog } from '../../../src/components';
 
 const meta: Meta<typeof PromptDialog> = {
     component: PromptDialog,
@@ -13,5 +12,5 @@ type Story = StoryObj<typeof PromptDialog>;
 
 // Props documentation story for PromptDialog
 export const PromptDialogProps: Story = {
-    render: (args: any) => <PromptDialog {...args} />,
+    render: (args: object) => <PromptDialog {...args} />,
 };
