@@ -1,7 +1,10 @@
-import React, { useState } from 'react';
-import { DragAndDrop, Card, CardBody } from '../../src/components';
+import { useState } from 'react';
+import { DragAndDrop, Card, CardBody, ORIENTATION } from '../../src/components';
 
-export default function DragAndDropDemo(props: any) {
+export default function DragAndDropDemo(props: {
+    orientation: ORIENTATION;
+    showIndicator: boolean;
+}) {
     const [items, setItems] = useState(['Delhi', 'Mumbai', 'Bangalore', 'Hyderabad', 'Pune']);
 
     const onDrop = (start: number, end: number) => {
@@ -9,9 +12,19 @@ export default function DragAndDropDemo(props: any) {
         const item = newItems.splice(start, 1);
         newItems.splice(end, 0, item[0]);
         setItems(newItems);
-    }
+    };
 
-    return (<DragAndDrop orientation={props.orientation} onDrop={onDrop} showIndicator={props.showIndicator}>
-        {items.map((item, index) => <Card key={item}><CardBody>{item}</CardBody></Card>)}
-    </DragAndDrop>)
+    return (
+        <DragAndDrop
+            orientation={props.orientation}
+            onDrop={onDrop}
+            showIndicator={props.showIndicator}
+        >
+            {items.map((item) => (
+                <Card key={item}>
+                    <CardBody>{item}</CardBody>
+                </Card>
+            ))}
+        </DragAndDrop>
+    );
 }
